p2 - troubleshooting cont. 4
I should probably be explaining more concerning what I am doing.
Something seems to be killing the rest of my code after "Posted on."
Anything after that just doesn't show up.  I know php is working
because all the other PHP works fine.

My guess is that it's the Time...Okay I just figured it out as I was
typing this.  Sigh, coding is 99% banging my head till I reach the
little blips of "ohhhh...sigh...that's why it's not working."

This error only happen(s)/(ed) on my LIVE server.  The LOCAL one is
fine, so I can't test it localy, well, not in anyway I can think of at
the moment.